requirements.txt
certifi==2019.9.11
dj-database-url==0.5.0
Django==2.2.7
django-crispy-forms==1.8.0
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.10.3
gunicorn==20.0.0
Markdown==3.1.1
pipenv==2018.11.26
python-dotenv==0.10.3
pytz==2019.3
sqlparse==0.3.0
virtualenv==16.7.7
virtualenv-clone==0.5.3
whitenoise==4.1.4
